Fix benchmarks and add Alonzo era to applyTx benchmarks
* Move a module with `HasTrace` instance to the test library for reuse in benchmarks

* Get rid of some dead code
