+++ eras/alonzo/test-suite/cardano-ledger-alonzo-test.cabal
@@ -43,12 +43,13 @@ library
   import:             base, project-config
 
   exposed-modules:
+    Test.Cardano.Ledger.Alonzo.AlonzoEraGen
     Test.Cardano.Ledger.Alonzo.EraMapping
     Test.Cardano.Ledger.Alonzo.Examples.Consensus
-    Test.Cardano.Ledger.Alonzo.Serialisation.Generators
-    Test.Cardano.Ledger.Alonzo.AlonzoEraGen
-    Test.Cardano.Ledger.Alonzo.Scripts
     Test.Cardano.Ledger.Alonzo.PlutusScripts
+    Test.Cardano.Ledger.Alonzo.Serialisation.Generators
+    Test.Cardano.Ledger.Alonzo.Scripts
+    Test.Cardano.Ledger.Alonzo.Trace
   build-depends:
     bytestring,
     cardano-binary,
@@ -67,8 +68,11 @@ library
     QuickCheck,
     cardano-ledger-shelley-test,
     cardano-ledger-shelley,
+    small-steps,
+    small-steps-test,
     strict-containers,
     text,
+    transformers,
   hs-source-dirs:
     src
 
@@ -87,8 +91,6 @@ test-suite cardano-ledger-alonzo-test
     Test.Cardano.Ledger.Alonzo.Serialisation.CDDL
     Test.Cardano.Ledger.Alonzo.Serialisation.Tripping
     Test.Cardano.Ledger.Alonzo.Translation
-    Test.Cardano.Ledger.Alonzo.Trials
-    Test.Cardano.Ledger.Alonzo.Trials
     Test.Cardano.Ledger.Alonzo.TxInfo
   build-depends:
     base16-bytestring,
@@ -99,14 +101,12 @@ test-suite cardano-ledger-alonzo-test
     cardano-ledger-alonzo-test,
     cardano-ledger-shelley-ma,
     cardano-ledger-core,
-    cardano-ledger-pretty,
     cardano-ledger-shelley-ma-test,
     cardano-protocol-tpraos,
     cardano-slotting,
     containers,
     data-default-class,
     plutus-core,
-    plutus-tx,
     plutus-ledger-api,
     QuickCheck,
     small-steps,
@@ -118,4 +118,3 @@ test-suite cardano-ledger-alonzo-test
     tasty-hunit,
     tasty-quickcheck,
     time,
-    transformers,
