+++ eras/shelley-ma/impl/cardano-ledger-shelley-ma.cabal
@@ -57,7 +57,6 @@ library
     cardano-crypto-class,
     cardano-data,
     cardano-ledger-core,
-    cardano-prelude,
     cardano-slotting,
     cborg,
     containers,
